+++ app/src/main/java/com/app/arcabyolimpo/data/remote/api/ArcaApi.kt
@@ -1,21 +1,26 @@
 package com.app.arcabyolimpo.data.remote.api
-
 import com.app.arcabyolimpo.data.remote.dto.auth.LoginRequestDto
 import com.app.arcabyolimpo.data.remote.dto.auth.LoginResponseDto
 import com.app.arcabyolimpo.data.remote.dto.auth.RefreshRequestDto
 import com.app.arcabyolimpo.data.remote.dto.auth.RefreshResponseDto
 import com.app.arcabyolimpo.data.remote.dto.beneficiaries.BeneficiariesListDto
 import com.app.arcabyolimpo.data.remote.dto.beneficiaries.BeneficiaryDto
+import com.app.arcabyolimpo.data.remote.dto.beneficiaries.GetBeneficiariesDisabilitiesDto
 import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
 import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordDto
 import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordDto
 import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.VerifyTokenResponseDto
-import com.app.arcabyolimpo.data.remote.dto.supplies.AcquisitionDto
 import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchDto
+import com.app.arcabyolimpo.data.remote.dto.product.ProductDto
+import com.app.arcabyolimpo.data.remote.dto.product.ProductRegisterInfoDto
+import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchModifyDto
+import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchRegisterDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.AcquisitionDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteResponseDto
+import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteSupplyBatchDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.GetFiltersDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.RegisterSupplyBatchDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.SuccessMessageDto
@@ -27,8 +32,11 @@ import com.app.arcabyolimpo.data.remote.dto.user.UserDto
 import com.app.arcabyolimpo.data.remote.dto.user.registeruser.RegisterResponseDto
 import com.app.arcabyolimpo.data.remote.dto.user.registeruser.RegisterUserDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.AddNewWorkshopDto
+import com.app.arcabyolimpo.data.remote.dto.workshops.DeleteResponseWorkshopDto
+import com.app.arcabyolimpo.data.remote.dto.workshops.DeleteWorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopDto
 import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopsListDto
+import com.app.arcabyolimpo.data.remote.dto.workshops.WorkshopResponseDto
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.Response
@@ -37,10 +45,10 @@ import retrofit2.http.DELETE
 import retrofit2.http.GET
 import retrofit2.http.Multipart
 import retrofit2.http.POST
+import retrofit2.http.PUT
 import retrofit2.http.Part
 import retrofit2.http.Path
 import retrofit2.http.Query
-
 
 /**
  * Defines the remote API endpoints.
@@ -58,6 +66,10 @@ interface ArcaApi {
     suspend fun refresh(
         @Body request: RefreshRequestDto,
     ): RefreshResponseDto
+
+    // External Collabs ----------------
+    //@GET("externalCollabs/")
+    //suspend fun getAllCollabs(): List<ExternalCollabDto>
 
     @GET("user/{id}")
     suspend fun getUserById(
@@ -127,16 +139,26 @@ interface ArcaApi {
     @GET("supplyBatch/acquisition/types")
     suspend fun getAcquisitionTypes(): List<AcquisitionDto>
 
-    @DELETE("supplyBatch/{id}")
+    @POST("supplyBatch/delete")
     suspend fun deleteSupplyBatch(
-        @Path("id") id: String,
-    )
+        @Body request: DeleteSupplyBatchDto,
+    ): DeleteResponseDto
 
-    // My route is a soft delete and an update
+    /**
+     * Deletes a single supply from the backend.
+     *
+     * This endpoint receives a [DeleteDto] with the information needed
+     * to identify which supply should be removed ( its ID).
+     *
+     * @param requestBody Data transfer object that contains the supply
+     * information required by the API to perform the delete operation.
+     * @return [DeleteResponseDto] containing the result of the delete
+     * operation, such as a success flag and/or a confirmation message.
+     */
     @POST("supplies/delete")
     suspend fun deleteOneSupply(
         @Body requestBody: DeleteDto,
-        // DeleteResponseDto es para la respuesta , para el snackbar
+        // DeleteResponseDto is for the response, for the snackbar
     ): DeleteResponseDto
 
     // Workshop ---------------------------
@@ -146,18 +168,37 @@ interface ArcaApi {
 
     @GET("workshop/search")
     suspend fun searchWorkshops(
-        @Query("nameWorkshop") name: String
+        @Query("nameWorkshop") name: String,
     ): List<WorkshopDto>
 
     @GET("workshop/{id}")
     suspend fun getWorkshop(
         @Path("id") id: String,
-    ): WorkshopDto
+    ): WorkshopResponseDto
 
     @POST("workshop/add")
     suspend fun addWorkshop(
         @Body requestBody: WorkshopDto,
     ): AddNewWorkshopDto
+
+    /**
+     * Deletes a single Workshop from the backend.
+     *
+     * This endpoint receives a [DeleteWorkshopDto] with the information needed
+     * to identify which supply should be removed ( its ID).
+     *
+     * @param requestBody Data transfer object that contains the supply
+     * information required by the API to perform the delete operation.
+     * @return [DeleteResponseWorkshopDto] containing the result of the delete
+     * operation, such as a success flag and/or a confirmation message.
+     */
+    @POST("workshop/delete")
+    suspend fun deleteWorkshops(
+        @Body requestBody: DeleteWorkshopDto,
+        // DeleteResponseDto is for the response, for the snackbar
+    ): DeleteResponseWorkshopDto
+
+
 
     // Beneficiary -------------
     @GET("beneficiary/list")
@@ -168,10 +209,26 @@ interface ArcaApi {
         @Path("id") id: String,
     ): BeneficiaryDto
 
+    /**
+     * Makes a soft delete to the selected beneficiary.
+     */
     @DELETE("beneficiary/{id}")
     suspend fun deleteBeneficiary(
         @Path("id") id: String,
     ): Response<Unit>
+
+    @GET("beneficiary/categories")
+    suspend fun getDisabilities(): GetBeneficiariesDisabilitiesDto
+
+    @POST("beneficiary/filter")
+    suspend fun filterBeneficiaries(
+        @Body params: FilterDto,
+    ): List<BeneficiaryDto>
+
+    @GET("beneficiary/search")
+    suspend fun searchBeneficiaries(
+        @Query("term") searchTerm: String,
+    ): List<BeneficiaryDto>
 
     @GET("supplies/workshop/category")
     suspend fun getWorkshopCategoryList(): WorkshopCategoryListDto
@@ -184,7 +241,7 @@ interface ArcaApi {
         @Part("unidadMedida") measureUnit: RequestBody,
         @Part("idCategoria") idCategory: RequestBody,
         @Part("status") status: RequestBody,
-        @Part imagenInsumo: MultipartBody.Part?
+        @Part imagenInsumo: MultipartBody.Part?,
     )
 
     @GET("productBatch/")
@@ -197,8 +254,20 @@ interface ArcaApi {
 
     @POST("productBatch/")
     suspend fun addProductBatch(
-        @Body batch: ProductBatchDto,
+        @Body batch: ProductBatchRegisterDto,
     )
+
+    @PUT("productBatch/{id}")
+    suspend fun modifyProductBatch(
+        @Path("id") id: String,
+        @Body batch: ProductBatchModifyDto,
+    )
+
+    @DELETE("productBatch/{id}")
+    suspend fun deleteProductBatch(
+        @Path("id") id: String,
+    )
+
     // Products --------------------------
 
     @Multipart
@@ -210,12 +279,38 @@ interface ArcaApi {
         @Part("idCategoria") idCategory: RequestBody,
         @Part("Descripcion") description: RequestBody,
         @Part("Disponible") status: RequestBody,
-        @Part image: MultipartBody.Part?
+        @Part image: MultipartBody.Part?,
     )
-
     @DELETE("product/{idProduct}")
     suspend fun deleteProduct(
         @Path("idProduct") idProduct: String,
     ): Response<Unit>
 
+    @GET("product/")
+    suspend fun getProducts(): List<ProductDto>
+
+    @GET("product/search")
+    suspend fun searchProducts(
+        @Query("q") query: String,
+    ): List<ProductDto>
+
+    @GET("product/add")
+    suspend fun getProductFilters(): ProductRegisterInfoDto
+
+    @GET("product/disponible")
+    suspend fun getProductsByAvailability(
+        @Query("disponible") disponible: Int,
+    ): List<ProductDto>
+
+
+    @GET("product/workshop")
+    suspend fun getProductsByWorkshop(
+        @Query("idTaller") idTaller: String,
+    ): List<ProductDto>
+
+    @GET("product/order")
+    suspend fun getProductsOrderedByPrice(
+        @Query("orderBy") orderBy: String,
+        @Query("direction") direction: String,
+    ): List<ProductDto>
 }