+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/supplies/SuppliesListDto.kt
@@ -3,10 +3,7 @@ package com.app.arcabyolimpo.data.remote.dto.supplies
 import com.google.gson.annotations.SerializedName
 
 data class SuppliesListDto(
-    @SerializedName("idInsumo")
-    val id: String,
-    @SerializedName("nombre")
-    val name: String,
-    @SerializedName("imagenInsumo")
-    val image: String?, // ← Hacer nullable porque puede ser null
+    @SerializedName("idInsumo") val id: String,
+    @SerializedName("nombre") val name: String,
+    @SerializedName("imagenInsumo") val image: String?
 )