+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/productbatches/productBatchModify/ProductBatchModifyUiState.kt
@@ -1,0 +1,9 @@
+package com.app.arcabyolimpo.presentation.screens.productbatches.productBatchModify
+
+data class ProductBatchModifyUiState(
+    val precioVenta: String = "",
+    val cantidadProducida: String = "",
+    val fechaCaducidad: String = "",
+    val fechaRealizacion: String = "",
+    val error: String? = null,
+)