+++ CHANGELOG.md
@@ -1,7 +1,37 @@
+## [0.1.0] - 17/11/2025
+
+### Added
+- Delete supply batch
+- Added filter and order products
+
+## [0.1.0] - 16/11/2025
+
+### Added
+ - Added products search
+ - Add DeleteWorkshop Logic, Data and Domain
+ - Add way to get users for workshops
+ - Added functionality search beneficiaries
+ - Added functionality view a Workshop (Video functionality needs more info)
+ - Filter and order beneficiaries functionality
+ - Fixed merge conflicts
+ - Added US to Delete Product Batches
+
+### Fixed
+- Refreshing on Product Batches List
+
+## [0.1.0] - 15/11/2025
+### Fixed
+
+
 ## [0.1.0] - 15/11/2025
 
 ### Added
 - Added function to the delete products button
+- Added US to Register Product Batches
+
+### Fixed
+- Fixed StandardInput missing parameter
+- Fixed error from standard input attibutes into register supply batch screen
 
 ## [0.1.0] - 14/11/2025
 
@@ -19,12 +49,14 @@
 - Added Workshop Card atom.
 - Added screen for view Workshop 
 - Added function for searching workshops by name.
+- Added US to Register Product Batches
 
 ### Fixed
 - ArcaApi get all external collabs
 - Search users
 - Added presentation, data and domain for RegisterSupplyBatch
 - Fix snackbar for register new workshops
+- Fix error of implementation in Input atoms
 
 ## [0.1.0] - 12/11/2025